Move IndexTree to util
